# Read message in buffer capacity first