The original message read involves at least two reads:
The first one is to read the message size. The second one is to read the rest of the message based on the message size that’s read in the first read. This change optimizes the incoming message read to attempt to read the full buffer capacity in the first read. It then decides whether or not subsequent reads are required based on `bytesRead` from the first read.

Also increase `INITIAL_BUFFER_SIZE` to `18000` from `10000` based on the average message size observed in #9708.

Implements #9711

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>