@a7ehuo 2020-06-11
As mentioned in https://github.com/eclipse/openj9/issues/9708#issuecomment-641540266, the latest average message size is ~17.5 KB, this change starts the experiment with `INITIAL_BUFFER_SIZE` as `18000`. 

`readCompileRequest()` is independent of `read()`, maybe we could start the experiment reading the message in full capacity for compilation request messages and the rest of messages will still use the default way to read messages (two reads: the first read is the size of the message; the second read, the message data). 

One potential issue with this approach, if we read the message in full capacity but there are still bytes left to read. A new buffer has to be allocated and the content from the old buffer with the full old capacity has to be copied into the new buffer. The current implementation might avoid the problem, it reads the size first and expands the buffer based on the size, at the time of getting a new buffer, the old buffer hasn’t been read into yet.