@a7ehuo 2020-06-16
Compared 76f5345 with the baseline (default read) running AcmeAir. no obvious performance gain so far. 

The logic in 76f5345 is as below:
```
- read the buffer once 
- check if the read across the message boundary, or if the bytes read is zero.
   - Yes, throw a StreamFailure exception.
   - No. check if more bytes to read based on the message size.
     -- No, done => deserialize.
     -- Yes, the buffer could be full or still space left. expand the buffer if the message size is greater than the buffer capacity. 
          -- Read all the rest of the messages.
          -- done => deserialize.
```

=== Baseline ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-message-size-default/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:verbose,dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:none -Xmx256m
Throughput	avg=11204.25	min=11119.10	max=11304.20	stdDev=60.4	maxVar=1.66%	confInt=0.31%	samples=10
Intermediate results:
Run 0	1235.9	11091.4	11193.7	Avg=11194	CPU=12289 ms  Footprint=235056 KB
Run 1	1194.6	11123.7	11132.7	Avg=11133	CPU=12993 ms  Footprint=233676 KB
Run 2	1211.8	11218.3	11256.4	Avg=11256	CPU=12783 ms  Footprint=238840 KB
Run 3	1202.4	11224.4	11270.7	Avg=11271	CPU=12586 ms  Footprint=241004 KB
Run 4	1212.2	11132.7	11225.1	Avg=11225	CPU=12711 ms  Footprint=233868 KB
Run 5	1178.5	11156.2	11119.1	Avg=11119	CPU=12508 ms  Footprint=233284 KB
Run 6	1189.7	11133.0	11189.7	Avg=11190	CPU=12728 ms  Footprint=240212 KB
Run 7	1184.7	11237.0	11304.2	Avg=11304	CPU=12299 ms  Footprint=239676 KB
Run 8	1214.2	11185.1	11197.3	Avg=11197	CPU=12419 ms  Footprint=233460 KB
Run 9	1203.8	11159.9	11153.6	Avg=11154	CPU=12547 ms  Footprint=233980 KB
CompTime	avg=12586.30	min=12289.00	max=12993.00	stdDev=222.7	maxVar=5.73%	confInt=1.03%	samples=10
Footprint	avg=236305.60	min=233284.00	max=241004.00	stdDev=3200.2	maxVar=3.31%	confInt=0.78%	samples=10
```

=== 76f5345 ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-message-size-newRead/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:verbose,dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:none -Xmx256m
Throughput	avg=11170.72	min=11014.10	max=11368.00	stdDev=124.6	maxVar=3.21%	confInt=0.65%	samples=10
Intermediate results:
Run 0	1210.1	11105.3	11145.1	Avg=11145	CPU=12632 ms  Footprint=232956 KB
Run 1	1178.3	11203.9	11290.8	Avg=11291	CPU=12614 ms  Footprint=233180 KB
Run 2	1176.4	11033.7	11093.1	Avg=11093	CPU=12555 ms  Footprint=231836 KB
Run 3	1209.1	10955.6	11014.1	Avg=11014	CPU=12758 ms  Footprint=233372 KB
Run 4	1193.6	11138.2	11368.0	Avg=11368	CPU=12617 ms  Footprint=238568 KB
Run 5	1192.1	11293.8	11307.5	Avg=11308	CPU=12583 ms  Footprint=239892 KB
Run 6	1183.1	11022.8	11047.9	Avg=11048	CPU=12675 ms  Footprint=232336 KB
Run 7	1188.5	11219.1	11229.5	Avg=11230	CPU=12225 ms  Footprint=232844 KB
Run 8	1186.8	11124.7	11174.2	Avg=11174	CPU=12474 ms  Footprint=233408 KB
Run 9	1164.7	11106.2	11037.0	Avg=11037	CPU=12450 ms  Footprint=234016 KB
CompTime	avg=12558.30	min=12225.00	max=12758.00	stdDev=147.6	maxVar=4.36%	confInt=0.68%	samples=10
Footprint	avg=234240.80	min=231836.00	max=239892.00	stdDev=2714.1	maxVar=3.47%	confInt=0.67%	samples=10
```