@a7ehuo 2020-06-16
76f5345 only apply the new read logic at the server side.