@a7ehuo 2020-06-17
6deb5f0 applies the new read() to both client and server streams and still don't see an improvement regarding to throughput. 

Baseline: Throughput  avg=`11159.40`
6deb5f0: Throughput  avg=`11079.16`

=== Baseline ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-message-size-default/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:none -Xmx256m
Throughput  avg=11159.40    min=10913.80    max=11309.30    stdDev=114.4    maxVar=3.62%    confInt=0.59%   samples=10
Intermediate results:
Run 0   1203.1  11275.1 11309.3 Avg=11309   CPU=12254 ms  Footprint=236572 KB
Run 1   1202.8  11165.4 11246.9 Avg=11247   CPU=12021 ms  Footprint=233376 KB
Run 2   1198.0  11114.1 11098.7 Avg=11099   CPU=12550 ms  Footprint=233416 KB
Run 3   1146.6  10958.0 10913.8 Avg=10914   CPU=12330 ms  Footprint=235540 KB
Run 4   1183.0  11129.4 11121.3 Avg=11121   CPU=11975 ms  Footprint=233000 KB
Run 5   1206.6  11183.3 11224.0 Avg=11224   CPU=12140 ms  Footprint=233884 KB
Run 6   1170.3  11025.1 11203.7 Avg=11204   CPU=11947 ms  Footprint=233296 KB
Run 7   1168.1  11121.1 11114.5 Avg=11114   CPU=11979 ms  Footprint=233772 KB
Run 8   1173.8  11211.9 11261.9 Avg=11262   CPU=11982 ms  Footprint=231400 KB
Run 9   1152.7  11095.6 11099.9 Avg=11100   CPU=12445 ms  Footprint=235128 KB
CompTime    avg=12162.30    min=11947.00    max=12550.00    stdDev=219.8    maxVar=5.05%    confInt=1.05%   samples=10
Footprint   avg=233938.40   min=231400.00   max=236572.00   stdDev=1463.4   maxVar=2.24%    confInt=0.36%   samples=10
```

=== 6deb5f0 ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-message-size-newRead/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:none -Xmx256m
Throughput  avg=11079.16    min=10865.50    max=11263.30    stdDev=125.3    maxVar=3.66%    confInt=0.66%   samples=10
Intermediate results:
Run 0   1179.3  11037.6 11055.6 Avg=11056   CPU=11898 ms  Footprint=233108 KB
Run 1   1202.2  11052.4 11150.5 Avg=11150   CPU=12514 ms  Footprint=232852 KB
Run 2   1181.6  10950.6 10931.6 Avg=10932   CPU=12220 ms  Footprint=233508 KB
Run 3   1187.0  11178.8 11149.5 Avg=11150   CPU=11706 ms  Footprint=233400 KB
Run 4   1158.9  10952.3 10865.5 Avg=10866   CPU=11851 ms  Footprint=233512 KB
Run 5   1164.1  11134.1 11182.2 Avg=11182   CPU=11803 ms  Footprint=233492 KB
Run 6   1173.1  11139.1 10955.7 Avg=10956   CPU=11804 ms  Footprint=233316 KB
Run 7   1198.9  11212.5 11263.3 Avg=11263   CPU=11652 ms  Footprint=232876 KB
Run 8   1165.4  11084.8 11131.5 Avg=11132   CPU=11758 ms  Footprint=238960 KB
Run 9   1170.6  11138.3 11106.2 Avg=11106   CPU=12155 ms  Footprint=235020 KB
CompTime    avg=11936.10    min=11652.00    max=12514.00    stdDev=273.2    maxVar=7.40%    confInt=1.33%   samples=10
Footprint   avg=234004.40   min=232852.00   max=238960.00   stdDev=1843.5   maxVar=2.62%    confInt=0.46%   samples=10
```