@mpirvu 2020-06-17
I don't expect this change to improve throughput, but rather CompTime which in turn can generate faster ramp-up when running in CPU constrained environments.
The numbers do show a small improvement 11.9/12.1=1.9%, so this is good!
The CI is about the same 1-1.3%, so it's hard to say exactly by how much we improved.