@a7ehuo 2020-06-17
With `20x` runs of AcmeAir, CompTime seems to improve a bit more: `~3.3%`.

|               | Throughput        | CompTime         |
--------- |---------------- |-----------------|
|Baseline | avg=`11069.01` | avg=`12301.65`|
|6deb5f0 | avg=`11108.06` | avg=`11896.45`|


=== Baseline ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-message-size-default/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:none -Xmx256m
Throughput	avg=11069.01	min=10813.40	max=11220.30	stdDev=95.4	maxVar=3.76%	confInt=0.33%	samples=20
Intermediate results:
Run 0	1181.7	11181.3	11116.8	Avg=11117	CPU=12266 ms  Footprint=233716 KB
Run 1	1179.5	11016.7	11043.7	Avg=11044	CPU=12296 ms  Footprint=232976 KB
Run 2	1179.3	11188.5	11099.4	Avg=11099	CPU=12100 ms  Footprint=232748 KB
Run 3	1192.0	11020.4	11134.4	Avg=11134	CPU=12068 ms  Footprint=233360 KB
Run 4	1162.6	11221.3	11147.1	Avg=11147	CPU=12242 ms  Footprint=233480 KB
Run 5	1168.1	11250.6	11124.8	Avg=11125	CPU=11940 ms  Footprint=240336 KB
Run 6	1189.3	11012.5	11111.5	Avg=11112	CPU=12345 ms  Footprint=233952 KB
Run 7	1169.1	10826.8	10813.4	Avg=10813	CPU=12499 ms  Footprint=235956 KB
Run 8	1171.3	11054.7	11131.0	Avg=11131	CPU=12402 ms  Footprint=232824 KB
Run 9	1181.5	11095.7	11166.2	Avg=11166	CPU=12498 ms  Footprint=233808 KB
Run 10	1156.0	10980.7	10947.4	Avg=10947	CPU=12536 ms  Footprint=234764 KB
Run 11	1124.6	11008.3	10961.0	Avg=10961	CPU=12314 ms  Footprint=233220 KB
Run 12	1178.4	10975.0	11006.3	Avg=11006	CPU=12310 ms  Footprint=232776 KB
Run 13	1158.0	10982.4	10944.1	Avg=10944	CPU=12385 ms  Footprint=232884 KB
Run 14	1141.2	11039.2	11133.9	Avg=11134	CPU=12352 ms  Footprint=242776 KB
Run 15	1196.8	11209.5	11220.3	Avg=11220	CPU=12314 ms  Footprint=240672 KB
Run 16	1164.4	11033.9	11050.0	Avg=11050	CPU=12328 ms  Footprint=233980 KB
Run 17	1163.6	11121.8	11054.0	Avg=11054	CPU=12373 ms  Footprint=233416 KB
Run 18	1194.5	11083.0	11109.2	Avg=11109	CPU=12312 ms  Footprint=233392 KB
Run 19	1170.9	11022.1	11065.7	Avg=11066	CPU=12153 ms  Footprint=233004 KB
CompTime	avg=12301.65	min=11940.00	max=12536.00	stdDev=147.5	maxVar=4.99%	confInt=0.46%	samples=20
Footprint	avg=234702.00	min=232748.00	max=242776.00	stdDev=2956.6	maxVar=4.31%	confInt=0.49%	samples=20
```


=== 6deb5f0 ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-message-size-newRead/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:none -Xmx256m
Throughput	avg=11108.06	min=10899.90	max=11274.50	stdDev=121.3	maxVar=3.44%	confInt=0.42%	samples=20
Intermediate results:
Run 0	1173.8	11031.2	10921.2	Avg=10921	CPU=11738 ms  Footprint=233012 KB
Run 1	1187.9	11079.5	11247.2	Avg=11247	CPU=11722 ms  Footprint=233540 KB
Run 2	1164.0	11038.8	10899.9	Avg=10900	CPU=12343 ms  Footprint=233924 KB
Run 3	1189.4	11091.3	11183.7	Avg=11184	CPU=11535 ms  Footprint=233480 KB
Run 4	1168.4	11086.2	11203.0	Avg=11203	CPU=11858 ms  Footprint=233084 KB
Run 5	1170.9	11114.1	11121.2	Avg=11121	CPU=11888 ms  Footprint=232868 KB
Run 6	1177.4	11008.7	10922.5	Avg=10922	CPU=11756 ms  Footprint=233128 KB
Run 7	1159.7	11271.4	11257.3	Avg=11257	CPU=12095 ms  Footprint=241568 KB
Run 8	1159.6	10976.9	11061.0	Avg=11061	CPU=11619 ms  Footprint=233084 KB
Run 9	1174.0	11032.6	11113.4	Avg=11113	CPU=11885 ms  Footprint=232732 KB
Run 10	1171.0	10924.7	10915.1	Avg=10915	CPU=11921 ms  Footprint=234548 KB
Run 11	1175.4	11130.1	11048.2	Avg=11048	CPU=11754 ms  Footprint=233496 KB
Run 12	1173.8	11088.7	11065.3	Avg=11065	CPU=11967 ms  Footprint=234272 KB
Run 13	1175.9	11035.5	11248.0	Avg=11248	CPU=11987 ms  Footprint=232956 KB
Run 14	1159.2	11162.3	11274.5	Avg=11274	CPU=12064 ms  Footprint=232376 KB
Run 15	1166.3	11090.0	11170.3	Avg=11170	CPU=12336 ms  Footprint=233604 KB
Run 16	1177.3	10985.9	11224.8	Avg=11225	CPU=11837 ms  Footprint=239184 KB
Run 17	1173.1	11106.0	11099.2	Avg=11099	CPU=11898 ms  Footprint=239308 KB
Run 18	1125.6	11036.1	11082.6	Avg=11083	CPU=11735 ms  Footprint=232884 KB
Run 19	1166.7	11017.2	11102.7	Avg=11103	CPU=11991 ms  Footprint=233340 KB
CompTime	avg=11896.45	min=11535.00	max=12343.00	stdDev=207.4	maxVar=7.00%	confInt=0.67%	samples=20
Footprint	avg=234319.40	min=232376.00	max=241568.00	stdDev=2547.1	maxVar=3.96%	confInt=0.42%	samples=20
```